Extract internal events into their own file
