+++ packages/xstate-wallet/src/workflows/application.ts
@@ -19,20 +19,24 @@ import * as CreateAndDirectFund from './create-and-direct-fund';
 import {sendDisplayMessage, MessagingServiceInterface, convertToChannelResult} from '../messaging';
 import {filter, map} from 'rxjs/operators';
 import * as CCC from './confirm-create-channel';
-import {
-  Participant,
-  UpdateChannelRequest,
-  CloseChannelRequest
-} from '@statechannels/client-api-schema';
+import {UpdateChannelRequest, CloseChannelRequest} from '@statechannels/client-api-schema';
 import {createMockGuard, getDataAndInvoke} from '../utils/workflow-utils';
 import {Store} from '../store/memory-store';
 import {StateVariables, SimpleAllocation} from '../store/types';
 import {ChannelStoreEntry} from '../store/memory-channel-storage';
-import {bigNumberify, BigNumber} from 'ethers/utils';
+import {bigNumberify} from 'ethers/utils';
 import * as ConcludeChannel from './conclude-channel';
 import {isSimpleEthAllocation} from '../utils/outcome';
 import {unreachable} from '../utils';
 import {deserializeAllocations} from '../serde/app-messages/deserialize';
+import {
+  PlayerRequestConclude,
+  PlayerStateUpdate,
+  ChannelUpdated,
+  JoinChannelEvent,
+  CreateChannelEvent,
+  WorkflowEvent
+} from '../event-types';
 
 export interface WorkflowContext {
   channelId?: string;
@@ -65,53 +69,7 @@ export interface WorkflowActions {
   updateStoreWithPlayerState: Action<WorkflowContext, PlayerStateUpdate>;
 }
 
-export interface JoinChannelEvent {
-  type: 'JOIN_CHANNEL';
-  channelId: string;
-  requestId: number;
-}
-// Events
-export type OpenEvent = CreateChannelEvent | JoinChannelEvent;
-
-export interface CreateChannelEvent {
-  type: 'CREATE_CHANNEL';
-  participants: Participant[];
-  outcome: SimpleAllocation;
-  appDefinition: string;
-  appData: string;
-  challengeDuration: BigNumber;
-  chainId: string;
-  requestId: number;
-}
-
-export interface ChannelUpdated {
-  type: 'CHANNEL_UPDATED';
-  storeEntry: ChannelStoreEntry;
-  requestId: number;
-}
-
-export interface PlayerStateUpdate {
-  type: 'PLAYER_STATE_UPDATE';
-  requestId: number;
-  outcome: SimpleAllocation;
-  channelId: string;
-  appData: string;
-}
-interface PlayerRequestConclude {
-  requestId: number;
-  type: 'PLAYER_REQUEST_CONCLUDE';
-  channelId: string;
-}
-type WorkflowEvent =
-  | PlayerRequestConclude
-  | PlayerStateUpdate
-  | OpenEvent
-  | ChannelUpdated
-  | JoinChannelEvent
-  | DoneInvokeEvent<string>;
-
-export type ApplicationWorkflowEvent = WorkflowEvent;
-
+export type ApplicationWorkflowEvent = WorkflowEvent | DoneInvokeEvent<string>;
 // TODO: Is this all that useful?
 export interface WorkflowServices extends Record<string, ServiceConfig<WorkflowContext>> {
   createChannel: (context: WorkflowContext, event: WorkflowEvent) => Promise<string>;
@@ -147,12 +105,17 @@ interface WorkflowStateSchema extends StateSchema<WorkflowContext> {
 }
 export type StateValue = keyof WorkflowStateSchema['states'];
 
-export type WorkflowState = State<WorkflowContext, WorkflowEvent, WorkflowStateSchema, any>;
+export type WorkflowState = State<
+  WorkflowContext,
+  ApplicationWorkflowEvent,
+  WorkflowStateSchema,
+  any
+>;
 
 const generateConfig = (
   actions: WorkflowActions,
   guards: WorkflowGuards
-): MachineConfig<WorkflowContext, WorkflowStateSchema, WorkflowEvent> => ({
+): MachineConfig<WorkflowContext, WorkflowStateSchema, ApplicationWorkflowEvent> => ({
   id: 'application-workflow',
   initial: 'initializing',
   on: {CHANNEL_UPDATED: {actions: [actions.sendChannelUpdatedNotification]}},
@@ -178,7 +141,7 @@ const generateConfig = (
         'getDataForCreateChannelConfirmation',
         'invokeCreateChannelConfirmation',
         'openChannelAndDirectFundProtocol'
-      ) as StateNodeConfig<WorkflowContext, {}, WorkflowEvent>),
+      ) as StateNodeConfig<WorkflowContext, {}, ApplicationWorkflowEvent>),
       entry: [actions.assignChannelId, actions.spawnObservers],
       exit: [actions.sendJoinChannelResponse]
     },
